# api/public.openapi.yaml
openapi: 3.0.3
info:
    title: WEBYCP Public API
    version: 0.7.0
    description: Public REST API used by the WEBYCP web application.
paths:
...
    /api/v1/auth/login:
        post:
            operationId: login
            summary: Create an authenticated session
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
            responses:
                "200":
                    description: Authenticated session.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/auth/logout:
        post:
            operationId: logout
            summary: Delete the current session
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "204":
                    description: Session deleted.
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/auth/me:
        get:
            operationId: getCurrentUser
            summary: Get the current authenticated user and CSRF token
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Current session.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
    /api/v1/auth/profile:
        patch:
            operationId: updateProfile
            summary: Update the authenticated user's profile and password
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateProfileRequest"
            responses:
                "200":
                    description: Updated authenticated session.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/accounts:
        get:
            operationId: listAccounts
            summary: List accessible hosting accounts
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Hosting accounts.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountListResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "500":
                    $ref: "#/components/responses/InternalError"
        post:
            operationId: createAccount
            summary: Create and queue provisioning for a hosting account
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateAccountRequest"
            responses:
                "202":
                    description: Account provisioning queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/accounts/{accountId}:
        patch:
            operationId: setAccount
            summary: Enable or disable a hosting account
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/AccountID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateEnabledRequest"
            responses:
                "202":
                    description: Account operation queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
        delete:
            operationId: deleteAccount
            summary: Delete an empty hosting account and quarantine its home
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/AccountID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Account deletion queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
    /api/v1/accounts/{accountId}/package:
        put:
            operationId: assignAccountPackage
            summary: Assign a Package to a hosting Account
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/AccountID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AssignAccountPackageRequest"
            responses:
                "200":
                    description: Updated Account Package assignment.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountOverview"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "422": { $ref: "#/components/responses/ValidationError" }
                "500": { $ref: "#/components/responses/InternalError" }
    /api/v1/packages:
        get:
            operationId: listPackages
            summary: List hosting Packages
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Hosting Packages.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PackageListResponse"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "500": { $ref: "#/components/responses/InternalError" }
        post:
            operationId: createPackage
            summary: Create a hosting Package
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WritePackageRequest"
            responses:
                "201":
                    description: Package created.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Package"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "422": { $ref: "#/components/responses/ValidationError" }
                "500": { $ref: "#/components/responses/InternalError" }
    /api/v1/packages/{packageId}:
        patch:
            operationId: updatePackage
            summary: Update a hosting Package
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/PackageID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WritePackageRequest"
            responses:
                "200":
                    description: Package updated.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Package"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "422": { $ref: "#/components/responses/ValidationError" }
                "500": { $ref: "#/components/responses/InternalError" }
        delete:
            operationId: deletePackage
            summary: Delete an unused hosting Package
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/PackageID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "204": { description: Package deleted. }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
    /api/v1/websites:
        get:
            operationId: listWebsites
            summary: List accessible websites
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Websites.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteListResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "500":
                    $ref: "#/components/responses/InternalError"
        post:
            operationId: createWebsite
            summary: Create and queue provisioning for a website
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateWebsiteRequest"
            responses:
                "202":
                    description: Website provisioning queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/websites/{websiteId}:
        patch:
            operationId: setWebsite
            summary: Update a website enabled state
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateWebsiteRequest"
            responses:
                "202":
                    description: Website operation queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
        delete:
            operationId: deleteWebsite
            summary: Delete a website and quarantine its document root
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Website deletion queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteJobResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/website-domains:
        get:
            operationId: listWebsiteDomains
            summary: List accessible website domains by kind
            security: [{ cookieAuth: [] }]
            parameters:
                - name: kind
                  in: query
                  required: true
                  schema: { type: string, enum: [primary, alias] }
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Website domains.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteDomainListResponse"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "422": { $ref: "#/components/responses/ValidationError" }
                "500": { $ref: "#/components/responses/InternalError" }
    /api/v1/websites/{websiteId}/domains:
        get:
            operationId: listWebsiteDomainsForWebsite
            summary: List hostname bindings for an accessible website
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteID"
            responses:
                "200":
                    description: Website domains.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteDomainCollection"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
        post:
            operationId: createWebsiteDomain
            summary: Add an alias domain to a website
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateWebsiteDomainRequest"
            responses:
                "202":
                    description: Website domain provisioning queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteDomainJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/website-domains/{websiteDomainId}:
        patch:
            operationId: setWebsiteDomain
            summary: Update a website domain hostname or enabled state
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteDomainID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateWebsiteDomainRequest"
            responses:
                "202":
                    description: Website domain operation queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteDomainJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
        delete:
            operationId: deleteWebsiteDomain
            summary: Delete an alias website domain
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/WebsiteDomainID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Website domain deletion queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WebsiteDomainJobResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/certificates:
        get:
            operationId: listCertificates
            summary: List accessible certificates
            security: [{ cookieAuth: [] }]
            parameters:
                - name: kind
                  in: query
                  schema: { type: string, enum: [website, panel] }
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Certificates.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CertificateListResponse"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "500": { $ref: "#/components/responses/InternalError" }
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: webycp_session
...
    responses:
        BadRequestError:
            description: The request is malformed.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        UnauthorizedError:
            description: Authentication is required.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
        ValidationError:
            description: The request is invalid.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        InternalError:
            description: An internal error occurred.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
    schemas:
...
        LoginRequest:
            type: object
            additionalProperties: false
            required: [username, password]
            properties:
                username:
                    type: string
                    minLength: 3
                    maxLength: 32
                password:
                    type: string
...
        AuthResponse:
            type: object
            additionalProperties: false
            required: [user, csrfToken, expiresAt, timezone]
            properties:
                user:
                    $ref: "#/components/schemas/User"
                csrfToken:
                    type: string
                expiresAt:
                    type: string
                    format: date-time
                timezone:
                    type: string
...